% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{can_be_integer}
\alias{can_be_integer}
\title{Check if a number could be considered to be an integer.}
\usage{
can_be_integer(x, na_rm = TRUE)
}
\arguments{
\item{x}{A numeric vector.}

\item{na_rm}{Should NAs be removed?}
}
\value{
A boolean. \code{TRUE} if the argument can be considered to be integer or
\code{FALSE} otherwise.
}
\description{
Could the input numeric vector be considered to be an integer vector, i.e. is
it equal to its floor?
}
\examples{
can_be_integer(c(3.0, 6))
can_be_integer(c(3.0, 6, NA))
can_be_integer(c(3, 3.5, NA))
}
