% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CanBeInteger}
\alias{CanBeInteger}
\title{Check if a number could be considered to be an integer.}
\usage{
CanBeInteger(x, na_rm = TRUE)
}
\arguments{
\item{x}{A numeric vector.}

\item{na_rm}{Should NAs be removed?}
}
\value{
A boolean. `TRUE` if the argument can be considered to be integer or
  `FALSE` otherwise.
}
\description{
After padding is removed, could the input numeric vector be considered to be
an integer vector, i.e. is it equal to its floor?
}
\examples{
CanBeInteger(c(3, 3.5, NA))
}
