% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_constructors.R
\name{masked_arr}
\alias{masked_arr}
\title{Masked array class.}
\usage{
masked_arr(arr, thresh)
}
\arguments{
\item{arr}{An array of logicals (the mask).}

\item{thresh}{The threshold. Either a scalar or an object of class \link{th}.}
}
\value{
An object of class \code{masked_arr}.
}
\description{
A \emph{mask} of an array with respect to a given threshold is found by taking the
original array and setting all elements falling below the threshold to
\code{FALSE} and the others to \code{TRUE}. An object of class \link{masked_arr} has the
attribute \code{thresh} detailing the threshold value that was applied.
}
